✅ ♻️ system-download.test.ts => split in multiple files & improve structure
